Streamline, consolidate de-clutter application composition - documentation still not
The application gets build progressively, originally each phase was built in separate notebook,  a complete application.
The last notebook created all the code of the previous notebooks, messy/impossible to
keep everything in sync. Changed over to single notebook (wikiImageAnalysis) that builds
each phase as an application and connects them with pub/sub. At the end of each phase get a view of the processing just added.

The videoImageAnalyis  notebook is derived from the wikiImageAnalysis for those how have video feeds to analyized.